requirements.txt
Django==2.0.6
gunicorn==19.8.1
psycopg2-binary==2.7.4
pytz==2018.4
djangorestframework==3.8.2
django-cors-headers==2.2.0
